+++ src/components/reusables/Input.tsx
@@ -3,7 +3,8 @@ import * as React from "react";
 export interface InputProps {
   type?: string;
   name: string;
-  label: string;
+  label?: string;
+  callback?: (e: any) => any;
 }
 
 export default class Input extends React.Component<InputProps, void> {
@@ -11,7 +12,7 @@ export default class Input extends React.Component<InputProps, void> {
     return(
       <div className="form-group">
         <label htmlFor={this.props.name}>{this.props.label}</label>
-        <input className="form-control" type={this.props.type || "text"} name={this.props.name} id={this.props.name} />
+        <input onChange={this.props.callback ? this.props.callback : null} ref="input" className="form-control" type={this.props.type || "text"} name={this.props.name} id={this.props.name} />
       </div>
     );
   };